chore: align about page lastmod
