+++ _headers
@@ -1,0 +1,7 @@
+/*
+  Strict-Transport-Security: max-age=31536000; includeSubDomains; preload
+  X-Content-Type-Options: nosniff
+  X-Frame-Options: SAMEORIGIN
+  Referrer-Policy: no-referrer-when-downgrade
+  Permissions-Policy: geolocation=(), microphone=(), camera=()
+  Content-Security-Policy-Report-Only: default-src 'self'; script-src 'self' 'unsafe-inline' https://cdn.jsdelivr.net; style-src 'self' 'unsafe-inline'; img-src 'self' data: https://telco.in https://www.telcoinnetwork.org https://www.telcoinassociation.org https://cdn.jsdelivr.net; font-src 'self'; connect-src 'self' https://www.telx.network https://telx.network https://api.telx.network; frame-ancestors 'none'
