+++ _redirects
@@ -1,0 +1,1 @@
+# Intentionally left blank; define redirects here if needed.
