+++ sitemap.xml
@@ -2,98 +2,98 @@
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
   <url>
     <loc>https://telcoinwiki.com/</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>weekly</changefreq>
     <priority>1.0</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/start-here.html</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>weekly</changefreq>
     <priority>0.9</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/faq/</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>weekly</changefreq>
     <priority>0.9</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/wallet.html</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>monthly</changefreq>
     <priority>0.8</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/digital-cash.html</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>monthly</changefreq>
     <priority>0.8</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/remittances.html</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>weekly</changefreq>
     <priority>0.8</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/tel-token.html</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>monthly</changefreq>
     <priority>0.8</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/network.html</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>monthly</changefreq>
     <priority>0.8</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/telx.html</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>monthly</changefreq>
     <priority>0.8</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/governance.html</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>monthly</changefreq>
     <priority>0.8</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/builders.html</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>monthly</changefreq>
     <priority>0.7</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/deep-dive.html</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>monthly</changefreq>
     <priority>0.7</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/links.html</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>monthly</changefreq>
     <priority>0.7</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/pools.html</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>weekly</changefreq>
     <priority>0.7</priority>
   </url>
   <url>
     <loc>https://telcoinwiki.com/portfolio.html</loc>
-    <lastmod>2025-09-17</lastmod>
+    <lastmod>2025-09-24</lastmod>
     <changefreq>weekly</changefreq>
     <priority>0.7</priority>
   </url>
-  <url>
-    <loc>https://telcoinwiki.com/about.html</loc>
-    <lastmod>2025-09-17</lastmod>
-    <changefreq>monthly</changefreq>
-    <priority>0.6</priority>
-  </url>
+    <url>
+      <loc>https://telcoinwiki.com/about.html</loc>
+      <lastmod>2025-09-24</lastmod>
+      <changefreq>monthly</changefreq>
+      <priority>0.6</priority>
+    </url>
 </urlset>
