+++ styles/index.css
@@ -1,0 +1,2 @@
+@import url('./variables.css');
+@import url('./site.css');
